requirements.txt
asgiref==3.7.2
Django==3.2.4
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
